This is one of the biggests updates I would have to say. I have finally been able to embed an iPython terminal into the application.
What this means is that users will no longer have to go outside of
the application to send direct commands to python! There is still
much work that will need to be done in order to get it to send
commands over the network but hey, it's all a work in progress.